+++ src/test/appdistro/client.spec.ts
@@ -8,10 +8,12 @@ import {
 import { FirebaseError } from "../../error";
 import * as api from "../../api";
 import * as nock from "nock";
+import { Distribution } from "../../appdistribution/distribution";
 
 describe("distribution", () => {
   const appId = "1:12345789:ios:abc123def456";
-  const distribution = new AppDistributionClient(appId);
+  const mockDistribution = new Distribution("./mockdata/mock.apk");
+  const appDistributionClient = new AppDistributionClient(appId);
 
   let sandbox: sinon.SinonSandbox;
 
@@ -29,58 +31,54 @@ describe("distribution", () => {
       nock(api.appDistributionOrigin)
         .get(`/v1alpha/apps/${appId}`)
         .reply(404, {});
-      return expect(distribution.getApp()).to.be.rejected;
+      return expect(appDistributionClient.getApp()).to.be.rejected;
     });
 
     it("should resolve when request succeeds", () => {
       nock(api.appDistributionOrigin)
         .get(`/v1alpha/apps/${appId}`)
         .reply(200, {});
-      return expect(distribution.getApp()).to.be.fulfilled;
+      return expect(appDistributionClient.getApp()).to.be.fulfilled;
     });
 
     it("should throw an error when the request fails", () => {
       nock(api.appDistributionOrigin)
         .get(`/v1alpha/apps/${appId}`)
         .reply(404, {});
-      return expect(distribution.getApp()).to.be.rejected;
+      return expect(appDistributionClient.getApp()).to.be.rejected;
     });
   });
 
-  describe("getJwtToken", () => {
-    it("should throw error if request fails", () => {
+  describe("uploadDistribution", () => {
+    it("should throw error if upload fails", () => {
       nock(api.appDistributionOrigin)
-        .get(`/v1alpha/apps/${appId}/jwt`)
+        .post(`/app-binary-uploads?app_id=${appId}`)
         .reply(400, {});
-      return expect(distribution.getJwtToken()).to.be.rejected;
+      return expect(appDistributionClient.uploadDistribution(mockDistribution)).to.be.rejected;
     });
 
-    describe("when request succeeds", () => {
-      it("should return null when response does not contain the token", () => {
-        nock(api.appDistributionOrigin)
-          .get(`/v1alpha/apps/${appId}/jwt`)
-          .reply(200, {});
-        return expect(distribution.getJwtToken()).to.be.eventually.eq(undefined);
-      });
-
-      it("should return the token", () => {
-        const fakeToken = "fake-token";
-        nock(api.appDistributionOrigin)
-          .get(`/v1alpha/apps/${appId}/jwt`)
-          .reply(200, { token: fakeToken });
-        return expect(distribution.getJwtToken()).to.be.eventually.eq(fakeToken);
-      });
+    it("should return token if upload succeeds", () => {
+      const fakeToken = "fake-token";
+      nock(api.appDistributionOrigin)
+        .post(`/app-binary-uploads?app_id=${appId}`)
+        .reply(200, { token: fakeToken });
+      return expect(appDistributionClient.uploadDistribution(mockDistribution)).to.be.eventually.eq(
+        fakeToken
+      );
     });
   });
 
   describe("pollReleaseIdByHash", () => {
     describe("when getUploadStatus returns IN_PROGRESS", () => {
       it("should throw error when retry count >= AppDistributionClient.MAX_POLLING_RETRIES", () => {
-        sandbox.stub(distribution, "getUploadStatus").resolves({
+        sandbox.stub(appDistributionClient, "getUploadStatus").resolves({
           status: UploadStatus.IN_PROGRESS,
         });
         return expect(
-          distribution.pollReleaseIdByHash("mock-hash", AppDistributionClient.MAX_POLLING_RETRIES)
+          appDistributionClient.pollUploadStatus(
+            "mock-hash",
+            AppDistributionClient.MAX_POLLING_RETRIES
+          )
         ).to.be.rejectedWith(
           FirebaseError,
           "it took longer than expected to process your binary, please try again"
@@ -90,14 +88,17 @@ describe("distribution", () => {
 
     it("should return release id when request succeeds", () => {
       const releaseId = "fake-release-id";
-      sandbox.stub(distribution, "getUploadStatus").resolves({
+      sandbox.stub(appDistributionClient, "getUploadStatus").resolves({
         status: UploadStatus.SUCCESS,
         release: {
           id: releaseId,
         },
       });
       return expect(
-        distribution.pollReleaseIdByHash("mock-hash", AppDistributionClient.MAX_POLLING_RETRIES)
+        appDistributionClient.pollUploadStatus(
+          "mock-hash",
+          AppDistributionClient.MAX_POLLING_RETRIES
+        )
       ).to.eventually.eq(releaseId);
     });
   });
@@ -109,7 +110,7 @@ describe("distribution", () => {
         .get(`/v1alpha/apps/${appId}/upload_status/${fakeHash}`)
         .reply(400, {});
 
-      return expect(distribution.getUploadStatus(fakeHash)).to.be.rejectedWith(
+      return expect(appDistributionClient.getUploadStatus(fakeHash)).to.be.rejectedWith(
         FirebaseError,
         "HTTP Error: 400"
       );
@@ -131,7 +132,9 @@ describe("distribution", () => {
           .get(`/v1alpha/apps/${appId}/upload_status/${fakeHash}`)
           .reply(200, response);
 
-        return expect(distribution.getUploadStatus(fakeHash)).to.eventually.deep.eq(response);
+        return expect(appDistributionClient.getUploadStatus(fakeHash)).to.eventually.deep.eq(
+          response
+        );
       });
     });
   });
@@ -139,7 +142,8 @@ describe("distribution", () => {
   describe("addReleaseNotes", () => {
     it("should return immediately when no release notes are specified", async () => {
       const apiSpy = sandbox.spy(api, "request");
-      await expect(distribution.addReleaseNotes("fake-release-id", "")).to.eventually.be.fulfilled;
+      await expect(appDistributionClient.addReleaseNotes("fake-release-id", "")).to.eventually.be
+        .fulfilled;
       expect(apiSpy).to.not.be.called;
     });
 
@@ -148,10 +152,9 @@ describe("distribution", () => {
       nock(api.appDistributionOrigin)
         .post(`/v1alpha/apps/${appId}/releases/${releaseId}/notes`)
         .reply(400, {});
-      return expect(distribution.addReleaseNotes(releaseId, "release notes")).to.be.rejectedWith(
-        FirebaseError,
-        "failed to add release notes"
-      );
+      return expect(
+        appDistributionClient.addReleaseNotes(releaseId, "release notes")
+      ).to.be.rejectedWith(FirebaseError, "failed to add release notes");
     });
 
     it("should resolve when request succeeds", () => {
@@ -159,15 +162,16 @@ describe("distribution", () => {
       nock(api.appDistributionOrigin)
         .post(`/v1alpha/apps/${appId}/releases/${releaseId}/notes`)
         .reply(200, {});
-      return expect(distribution.addReleaseNotes(releaseId, "release notes")).to.eventually.be
-        .fulfilled;
+      return expect(appDistributionClient.addReleaseNotes(releaseId, "release notes")).to.eventually
+        .be.fulfilled;
     });
   });
 
   describe("enableAccess", () => {
     it("should return immediately when testers and groups are empty", async () => {
       const apiSpy = sandbox.spy(api, "request");
-      await expect(distribution.enableAccess("fake-release-id")).to.eventually.be.fulfilled;
+      await expect(appDistributionClient.enableAccess("fake-release-id")).to.eventually.be
+        .fulfilled;
       expect(apiSpy).to.not.be.called;
     });
 
@@ -176,7 +180,8 @@ describe("distribution", () => {
       nock(api.appDistributionOrigin)
         .post(`/v1alpha/apps/${appId}/releases/${releaseId}/enable_access`)
         .reply(200, {});
-      return expect(distribution.enableAccess(releaseId, ["tester1"], ["group1"])).to.be.fulfilled;
+      return expect(appDistributionClient.enableAccess(releaseId, ["tester1"], ["group1"])).to.be
+        .fulfilled;
     });
 
     describe("when request fails", () => {
@@ -195,10 +200,9 @@ describe("distribution", () => {
             groupIds: groups,
           })
           .reply(412, { error: { status: "FAILED_PRECONDITION" } });
-        return expect(distribution.enableAccess(releaseId, testers, groups)).to.be.rejectedWith(
-          FirebaseError,
-          "failed to add testers/groups: invalid testers"
-        );
+        return expect(
+          appDistributionClient.enableAccess(releaseId, testers, groups)
+        ).to.be.rejectedWith(FirebaseError, "failed to add testers/groups: invalid testers");
       });
 
       it("should throw invalid groups error when status code is INVALID_ARGUMENT", () => {
@@ -209,10 +213,9 @@ describe("distribution", () => {
             groupIds: groups,
           })
           .reply(412, { error: { status: "INVALID_ARGUMENT" } });
-        return expect(distribution.enableAccess(releaseId, testers, groups)).to.be.rejectedWith(
-          FirebaseError,
-          "failed to add testers/groups: invalid groups"
-        );
+        return expect(
+          appDistributionClient.enableAccess(releaseId, testers, groups)
+        ).to.be.rejectedWith(FirebaseError, "failed to add testers/groups: invalid groups");
       });
 
       it("should throw default error", () => {
@@ -224,7 +227,7 @@ describe("distribution", () => {
           })
           .reply(400, {});
         return expect(
-          distribution.enableAccess(releaseId, ["tester1"], ["group1"])
+          appDistributionClient.enableAccess(releaseId, ["tester1"], ["group1"])
         ).to.be.rejectedWith(FirebaseError, "failed to add testers/groups");
       });
     });