+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 - Fixes issue where custom emulator ports were strings not numbers (#2287).
 - Fixes issue where Emulator UI was never downloaded during `firebase init emulators`.
-- Skip app selection in `firebase apps:sdkconfig` if there is a single matching app.
+- Skips app selection in `firebase apps:sdkconfig` if there is a single matching app.
 - Fixes incorrect warning about emulator configuration (#2290).
 - Improves error messages for `firebase init` with new projects (#2266).
 - Improves handling of permission denied errors during API enablement.
+- Removes deep proxying of `functions.config()` inside the Functions emulator.
+- Updates App Distribution upload URL.